Изменяет логику отрисовки компонентов
